+++ Labs/Lab 1/gameOfLife.h
@@ -22,25 +22,25 @@ Function to initialize all cells in the board.
 Takes in 2D array of pointers and size of array
 Should create structs and populate the array
 */
-void initCells(Cell* board[][10], int boardSize);
+void initCells(Cell* board[][10], size_t boardSize);
 
 /*
 Function to read the board from a file
 Prompt for the file to read inside of the function
 Structure of file should consist of 10 lines of 0 or 1 to indicate cell state
 */
-void readBoard(Cell* board[][10], int boardSize);
+void readBoard(Cell* board[][10], size_t boardSize);
 
 /*
 Function to print out all cells to cout
 */
-void printCells(Cell* board[][10], int boardSize);
+void printCells(Cell* board[][10], size_t boardSize);
 
 /*
 Function to count the number of live neighbors for each cell.
 Must use the x, y position stored with each cell to determine which neighbors they have
 */
-void findNumNeighbors(Cell* board[][10], int boardSize, Cell* curCell);
+void findNumNeighbors(Cell* board[][10], size_t boardSize, Cell* curCell);
 
 /*
 Function to update each cell's state based on number of neighbors
@@ -53,4 +53,4 @@ Any dead cell with exactly three live neighbors becomes a live cell, as if by re
 
 Return if you updated cells or not to break out of while loop from main.
 */
-bool updateCellState(Cell* board[][10], int boardSize);
+bool updateCellState(Cell* board[][10], size_t boardSize);
